feat(factory): define work board lifecycle
Move the built-in Work board topology and phase behavior into the typed board definition contract so lifecycle validation is shared across built-in boards.\n\nCo-Authored-By: Mastra Code (openai/gpt-5.6-sol) <noreply@mastra.ai>
