+++ .changeset/fruity-eggs-march.md
@@ -1,0 +1,11 @@
+---
+'@mastra/factory': minor
+---
+
+Added the built-in Work board definition so its phases and phase behavior use the typed board lifecycle API alongside the Review board.
+
+```ts
+import { workBoard } from '@mastra/factory';
+
+workBoard.allowsTransition('planning', 'execute');
+```
