+++ mastracode/factory/src/boards/define-board.test.ts
@@ -1,6 +1,7 @@
 import { describe, expect, it } from 'vitest';
 import { BoardDefinitionError, defineBoard } from './define-board.js';
 import { reviewBoard } from './review.js';
+import { workBoard } from './work.js';
 
 describe('defineBoard', () => {
   it('normalizes linear and outcome transitions', () => {
@@ -98,5 +99,20 @@ describe('defineBoard', () => {
     expect(reviewBoard.allowsTransition('canceled', 'review')).toBe(true);
     expect(reviewBoard.rules.intake?.pullRequest?.onEnter).toBeTypeOf('function');
     expect(reviewBoard.rules.review?.pullRequest?.onEnter).toBeTypeOf('function');
+  });
+
+  it('defines the built-in Work lifecycle and phase behavior', () => {
+    expect(workBoard.initialPhase).toBe('intake');
+    const phases = ['intake', 'triage', 'planning', 'execute', 'review', 'done', 'canceled'] as const;
+    for (const from of phases) {
+      for (const to of phases) {
+        expect(workBoard.allowsTransition(from, to), `${from} -> ${to}`).toBe(true);
+      }
+    }
+    expect(workBoard.rules.intake?.issue?.onEnter).toBeTypeOf('function');
+    expect(workBoard.rules.triage?.linearIssue?.onEnter).toBeTypeOf('function');
+    expect(workBoard.rules.planning?.manual?.onEnter).toBeTypeOf('function');
+    expect(workBoard.rules.execute?.issue?.onEnter).toBeTypeOf('function');
+    expect(workBoard.rules.done?.issue?.onEnter).toBeTypeOf('function');
   });
 });
