+++ mastracode/factory/src/boards/work.ts
@@ -1,0 +1,176 @@
+import type { FactoryRuleItemContext, FactoryStageRuleContext } from '../rules/types.js';
+import { needsApproval } from '../rules/types.js';
+import { workItemNumber } from '../work-item-branch.js';
+import { defineBoard } from './define-board.js';
+import type { BoardPhaseDefinition } from './define-board.js';
+
+function linearIdentifier(item: FactoryRuleItemContext): string | undefined {
+  const identifier = item.metadata?.identifier;
+  return typeof identifier === 'string' ? identifier : undefined;
+}
+
+function sourceRef(item: FactoryRuleItemContext): string {
+  const link = item.url ? ` (${item.url})` : '';
+  if (item.source === 'linear-issue') {
+    const identifier = linearIdentifier(item);
+    return identifier ? `Linear issue ${identifier}${link}` : `Linear issue ${item.title}${link}`;
+  }
+  if (item.source === 'manual') return item.url ? `Work item${link}` : item.title;
+  const noun = item.source === 'github-pr' ? 'GitHub pull request' : 'GitHub issue';
+  const number = workItemNumber(item);
+  if (number === undefined) return item.url ? `${noun}${link}` : item.title;
+  return `${noun} #${number}${link}`;
+}
+
+function invokeIssueInvestigation(context: FactoryStageRuleContext) {
+  return {
+    type: 'invokeSkill',
+    idempotencyKey: `${context.ingress.id}:factory-triage`,
+    role: 'triage',
+    skillName: 'factory-triage',
+    arguments: sourceRef(context.item),
+  } as const;
+}
+
+function prepareApproval(context: FactoryStageRuleContext) {
+  return {
+    type: 'invokeSkill',
+    idempotencyKey: `${context.ingress.id}:prepare-approval`,
+    role: 'triage',
+    prompt:
+      `Prepare approval for ${sourceRef(context.item)}. Review the existing triage comment and summarize ` +
+      'the decision needed before implementation or closure.',
+  } as const;
+}
+
+function triageIssueEntry(context: FactoryStageRuleContext) {
+  return needsApproval(context.item) ? prepareApproval(context) : invokeIssueInvestigation(context);
+}
+
+const LINEAR_FETCH_HINT =
+  "Start by fetching the issue's full details (description and comments) with the linear_get_issue tool.";
+
+function investigateTriagedLinearIssue(context: FactoryStageRuleContext) {
+  return {
+    type: 'invokeSkill',
+    idempotencyKey: `${context.ingress.id}:factory-triage-linear`,
+    role: 'triage',
+    skillName: 'factory-triage',
+    arguments: `${sourceRef(context.item)}\n\n${LINEAR_FETCH_HINT}`,
+  } as const;
+}
+
+function planWorkItem(context: FactoryStageRuleContext) {
+  return {
+    type: 'invokeSkill',
+    idempotencyKey: `${context.ingress.id}:factory-plan`,
+    role: 'plan',
+    skillName: 'factory-plan',
+    arguments: context.item.url ? `Work item (${context.item.url})` : context.item.title,
+  } as const;
+}
+
+const GITHUB_LOGIN = /^[a-zA-Z0-9](?:[a-zA-Z0-9]|-(?=[a-zA-Z0-9])){0,38}$/;
+
+function reporterCoAuthor(context: FactoryStageRuleContext) {
+  if (context.source !== 'issue') return undefined;
+  const author = context.item.metadata?.author;
+  if (typeof author !== 'string' || !author) return undefined;
+  if (author.endsWith('[bot]') || !GITHUB_LOGIN.test(author)) return undefined;
+  return author;
+}
+
+function buildWorkItem(context: FactoryStageRuleContext) {
+  const ref = sourceRef(context.item);
+  const fromApprovedPlan = context.fromStage === 'planning';
+  const task = fromApprovedPlan
+    ? `Implement the approved plan for ${ref}. Open a pull request when the work is ready for review.`
+    : `Implement a fix for ${ref}: investigate the root cause, make the change with tests, and open a pull request.`;
+  const reporter = reporterCoAuthor(context);
+  const credit = reporter
+    ? ` The work was reported by @${reporter}: credit them on every commit with a ` +
+      `\`Co-Authored-By: ${reporter} <ID+${reporter}@users.noreply.github.com>\` trailer, ` +
+      `resolving ID with \`gh api users/${reporter} --jq .id\`.`
+    : '';
+  return {
+    type: 'invokeSkill',
+    idempotencyKey: `${context.ingress.id}:build`,
+    role: 'work',
+    prompt: `${task}${credit}`,
+  } as const;
+}
+
+function completeIssue(context: FactoryStageRuleContext) {
+  return {
+    type: 'invokeSkill',
+    idempotencyKey: `${context.ingress.id}:factory-complete-issue`,
+    role: 'triage',
+    skillName: 'factory-complete-issue',
+    arguments: context.item.url ? `GitHub issue (${context.item.url})` : context.item.title,
+  } as const;
+}
+
+function onArrival<Effect>(rule: (context: FactoryStageRuleContext) => Effect) {
+  return (context: FactoryStageRuleContext): Effect | undefined => {
+    if (context.cause !== 'linked_item_materialized') return;
+    if (context.item.metadata?.autoStartCandidate !== true) return;
+    return rule(context);
+  };
+}
+
+export type WorkBoardPhase = 'intake' | 'triage' | 'planning' | 'execute' | 'review' | 'done' | 'canceled';
+
+const allOtherPhases = {
+  intake: 'intake',
+  triage: 'triage',
+  planning: 'planning',
+  execute: 'execute',
+  review: 'review',
+  done: 'done',
+  canceled: 'canceled',
+} as const;
+
+export const workBoard = defineBoard<'work', Record<WorkBoardPhase, BoardPhaseDefinition<WorkBoardPhase>>>({
+  id: 'work',
+  title: 'Work',
+  initialPhase: 'intake',
+  phases: {
+    intake: {
+      title: 'Intake',
+      outcomes: allOtherPhases,
+      onEnter: { issue: onArrival(triageIssueEntry) },
+    },
+    triage: {
+      title: 'Triage',
+      outcomes: allOtherPhases,
+      onEnter: { issue: triageIssueEntry, linearIssue: investigateTriagedLinearIssue },
+    },
+    planning: {
+      title: 'Planning',
+      outcomes: allOtherPhases,
+      onEnter: { issue: planWorkItem, linearIssue: planWorkItem, manual: planWorkItem },
+    },
+    execute: {
+      title: 'Building',
+      outcomes: allOtherPhases,
+      onEnter: { issue: buildWorkItem, linearIssue: buildWorkItem, manual: buildWorkItem },
+    },
+    review: {
+      title: 'Review',
+      outcomes: allOtherPhases,
+    },
+    done: {
+      title: 'Done',
+      outcomes: allOtherPhases,
+      onEnter: { issue: completeIssue },
+    },
+    canceled: {
+      title: 'Canceled',
+      outcomes: allOtherPhases,
+    },
+  },
+});
+
+export function isWorkBoardPhase(value: string): value is WorkBoardPhase {
+  return value in workBoard.phases;
+}
